second commit

+++ src/main/java/com/kgisl/sb10/entity/Person.java
@@ -2,5 +2,5 @@ package com.kgisl.sb10.entity;
 
 import org.springframework.data.annotation.Id;
  
-public record Person(@Id Long id, String uname, String email) {
+public record Person(@Id int id, String uname, String email) {
 }
