+++ src/main/java/com/kgisl/sb10/repository/PersonRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.ListCrudRepository;
 
 import com.kgisl.sb10.entity.Person;
 
-public interface PersonRepository extends ListCrudRepository<Person, Long> {
+public interface PersonRepository extends ListCrudRepository<Person, Integer> {
     
 }
